Update components for ability to listen to groups
A group is a property on a component that can be used for determining whether to act on an event that the component is listening for. This is useful when there are multiple instances of a component on the page (ex. products components). All non UI components are affected by this.

Products has been split into 3 components. ProductsBase, is the component to be extended from and does not listen for events. ProductsSayt listens for sayt products responses. Products listens for search responses.