SFX-354: Update Autocomplete and Products to listen to relevant events  (#34)
* Update components for ability to listen to groups

A group is a reflected property on a component that can be used for determining whether to act on an event that the component is listening for. This is useful when there are multiple instances of a component on the page (ex. products components). All non UI components are affected by this. The property defaults to an empty string at component construction and for all falsey values during comparisons.

Products has been split into 3 components. ProductsBase, is the component to be extended from and does not listen for events. ProductsSayt listens for SAYT_PRODUCTS_RESPONSE. Products listens for SEARCH_RESPONSE.