SFX-385: Allow for passing custom product transformer (#36)
Update the demo to define functions for transforming the product for Search and Sayt, much as the client would have to do. The transformer for the Apparel endpoint, previously hardcoded in the Logic Layer, is now defined in `demo.html`.

Note that the product transformer for Search is intentionally simple. It will display only product titles.

Additionally:
* Update Product and Searchbox components to extend `LitElement` instead of `Base`. This removes issues with multiple renders/updates of components (artifacts from previous renders remain in the DOM).
* Add event listeners for events in the demo file. This is intentional. It is easier to demonstrate functionality, and to debug issues, if we can see events being logged to the console. We can consider adding more event listeners in the future, or make a plugin that listens to all of our events and logs them out.
* Update `sfx-all-plugins-bundle.js`. This should likely be removed in the future.