ELE-480: Add data initialization check to cache requests (#56)
An _initialized boolean property was added to Base with the intention of being set to true once a component has received data either through a search call or attribute value. This value is checked to decide whether to request the cache and whether to use the cache data when reacting to cache responses.

An accompanying decorator, dataInitializer, was added to extend a property's set accessor to also set a component's _initialized value to true. The decorator is designed to be used on whatever property a cache response would populate.